requirements.txt
python-telegram-bot==20.8
python-dateutil==2.8.2
pytz==2024.1
python-dotenv==1.0.0